requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
djangorestframework-api-key==2.*
gunicorn
numpy
sendgrid
supabase
requests
beautifulsoup4
